The GBMPBL scheme had problems with occasional division by zero and accessing an out-of-bounds array element.
The division by zero was happening on the following line, when NSQUAR(K) was equal to zero: 

 908           BBLS(K) = MIN(rstbl*SQRT(TKE(K)/NSQUAR(K)),karman*ZQX(K))


We avoid this by making sure that when NSQUAR(K) is zero, a different code block is executed.   
It was a matter of changing a "less than zero" test to a "less than or equal to zero" test.  



Also,  there was an if-test that would access an invalid array element in the original code, namely kbot(ilay-1) 
when (ilay <= 1).   To prevent this, we simply broke the test into two parts: 

-          if(ilay.gt.1.and.ktop(ilay).eq.kbot(ilay-1))then 

becomes 

+          if(ilay.gt.1) then
+             if(ktop(ilay).eq.kbot(ilay-1))then 



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@6980 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d