The runtime error checks (from "configure -D") uncovered an unitialized array "q_ref2m" in the surface driver. This array only gets used when the CLM land surface model is turned on.
As it turns out, the garbage-filled contents of this array were not adversely affecting the output from WRF, 
since it never appears on the right-hand-side of an assignment statement inside of the CLM scheme.    
However, it is necessary to initialize this array to something in order for runtime error-checking to fully 
check the CLM scheme for "correctness".  



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@6981 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d