Add a namelist option to the time_control section of namelist.input to switch NetCDF to write in NOFILL mode instead of the default FILL mode. Background is that Jim Abeles at IBM discovered that WRF running at NCEP was generating output traffic equivalent to 2x the size of the write along with considerable input traffic when writing history and restarts through NetCDF.  A little more digging showed that this was the so-called FILL behavior, and it's the default NETCDF mode.  But there's also a NOFILL option that can be set using the nf_set_fill function in NetCDF on a per-file basis.  We check with UniData and they agree that using NOFILL for WRF history and restart writes is appropriate and safe. So the proposed change is to add a namelist variable, ncd_nofill, that can be set either .TRUE. or .FALSE.   The default will be .FALSE., meaning no change from current behavior.  But users such as NCEP and others (AFWA and ESRL might also benefit)  will be able to switch to NOFILL mode by setting ncd_nofill to .TRUE.
Jim observed that when writing in NOFILL mode, the performance of
NetCDF output is close to that of binary, at least for the cases he was
looking at.  But as a general rule, the NOFILL option will not have much
effect on the run time of a given run of WRF if quilting is also in use.
The main advantage is that it cuts by half the amount of traffic WRF
generates to the I/O system.  This could make a big difference for
overall performance of heavily loaded systemslike Jet or WCOSS.

Note: Unidata advises that the purpose of FILL mode is to enable integrity
checking on NetCDF files being accessed both read/write.  Therefore, for
safety, any WRF application that expects to both read and write to the
same open NetCDF file (does var do any of that?) ncd_nofill should be set
to .false. .  That's the main reason for having .false. as the default.

Affected files:

M    external/io_netcdf/wrf_io.F90
M    frame/module_io.F
M    Registry/registry.io_boilerplate




git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@7219 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d