Stolen from John M's original commit messsge:
"Add a namelist option to the time_control section of namelist.input
to switch NetCDF to write in NOFILL mode instead of the default FILL
mode. Background is that Jim Abeles at IBM discovered that WRF running at
NCEP was generating output traffic equivalent to 2x the size of the write
along with considerable input traffic when writing history and restarts
through NetCDF.  A little more digging showed that this was the so-called
FILL behavior, and it's the default NETCDF mode.  But there's
also a NOFILL option that can be set using the nf_set_fill function
in NetCDF on a per-file basis.  We check with UniData and they agree that
using NOFILL for WRF history and restart writes is appropriate and safe."

The default setting is now NOFILL=TRUE.  The stakeholders of the 
WRF infrastructure (NMM, Chem, DA) and the Unidata developer's group have 
all given their approval.

M    registry.io_boilerplate


git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@7335 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d