Brian Reen (with a careful eye), noticed that the usage of the option for ARW real (lowest_lev_from_surface) was broken.
The correct implementation is a 1-line change to modify the
vertical level in an assignment.  After all of the vertical 
interpolation for a variable is complete, if the setting for
lowest_lev_from_surface is true, then the lowest level (index
of k=1) for the input data from metgrid is assigned to the 
first eta level in the output of real (input to WRF).

Very few users activate this switch.  However, it will
modify the lowest level of the model input (and if those 
changes are close to a lateral boundary, the LBC file
as well).  The good news is that the changed behavior is
now correct (expected) behavior.

M    dyn_em/module_initialize_real.F



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@7336 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d