From an IBM power6 user, we got a small tweak (qstrict).  Probably the xlf compiler was being a bit too wild with some looping assumptions.
Also the user provides a Linux stanza with the xlf/xlc compilers.  This 
is likely not a widely used option, so it is pretty specific for this one 
user (the CPP has an env variable, for example).  However, it is a nice 
template if anyone else needs this capability.

M    configure_new.defaults



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@7541 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d