Mods for faster lateral boundary access for REALLY long climate cases had a bug: if the restart did not start at a boundary time, the code hung.
Now, for a restart, the WRF code finds the lateral boundary
time that is either equal to the restart, or the first 
lateral boundary time PAST the restart requested time.

If the LBC and the restart are at the same time, we are good.

If the LBC is just one time period past the restart time, 
then we back up the LBC file a single time period, and then we 
are good to go.

Also, in one of the author's opinions, the informational output 
messages are much nicer now.

Again, for finding this bug, nice catch Wei.

M    input_wrf.F



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@7612 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d