CORRECTED:
More changes to the configure script based on feedback from the developers on the last round of changes.

In configure_new.defaults, previously we simply printed the hand-edited description of the configuration option. This is not ideal, especially since many of these descriptions don't explicitly list the compiler names. I have modified configure_new.defaults to add a new variable to each option, named "DESCRIPTION", where all the human-readable information can be stored. The options are in this format:

COMPILERNAME ($SFC/$SCC): Extra information

"COMPILERNAME" is the generic name of the compiler, such as "INTEL", "PGI", etc. $SFC/$SCC are the actual names of the specified compilers, and are automatically substituted in from the values listed in the configuration option. "Extra information" is the extra info needed to disambiguate the similar options, such as SGI MPT, PGI accelerator, etc.

Also, the OS and Machine options are now listed in the "Please select" line; these are the values that the configure script finds with the `uname` and `uname -m` commands respectively.


Additionally, some minor changes:

1. Replacing tabs with spaces in configure_new.defaults. The script should work with any whitespace characters, but the current usage is inconsistent, so I went with spaces.

2. Add a check at the end of Config_new.pl to ensure that the user-selected option is found. If someone edits the script they may accidentally cause this check to fail (as I did), which led to a malformed configure.wrf file and other very confusing problems that can be avoided with a simple if statement.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@7738 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d