Add in new horizontal interpolation options (beyond only SINT) for ARW.  These include nearest neighbor (just for testing purposes), bilinear, and overlapping quadratic (yes, that is what was around in MM4 days).  These are available if someone needs to do UNRECOMMENDEDLY large nest ratio.
This does not need to be an advertised option, because it might 
encourage additional 15:1 (and beyond) ratio behaviors.  

Additionally, the lateral boundaries for the FG forcing, are 
pulled from the complete 2d array of interpolated data (for 
the test cases, this cost about +- 1% more per time step for 
a single processor run, which would be most sensitive). 

The default behavior is bit-wise identical results and functioning.

M       Registry/Registry.EM_COMMON
M       run/README.namelist
M       share/interp_fcn.F
M       share/mediation_integrate.F

This is controlled by the name list option, with the default given as:
&domains
 interp_method_type = 2

The available options are 
1: bilinear
2: SINT
3: nearest neighbor
4: overlalpping quadratic
12: SINT, but using the same lateral boundary framework as options 1, 
3, and 4.  Used as a test to insure the proper functioning of the 
lateral boundary routine.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@7741 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d