Put in more easy things for CF compliance: the date information. The string in the Registry (ARW only) has a special substring inserted that will be replaced by the WRF starting date.  That starting date allows us to say "minutes since YYYY-MM-DD HH:mm:ss", and THAT gets us time compliancy for CF.
A few other coordinate items are "in" but not used.  The trouble with
those is that CF seems to restrict the number of dimensions for
some of the coordinate variables.  For example, while it is great that
moving from 1 dimension to 2 dimensions for lat and lon is a huge win 
(now there can be a 2d array), the WRF model requires three dimensions
(we can have time).  Even for non-moving domains, all the output from
WRF has a time stamp.  So, this capability is not used until CF allows
more general coordinates, then we can revisit and see if this will
work out OK.  When trying to put the third dimension for lat and lon,
the "map" capability in ncview suddenly stopped working.

M    Registry/Registry.EM_COMMON
M    share/output_wrf.F
M    share/wrf_ext_write_field.F




git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@7872 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d