1.  performance mods for the two new RRTMG routines that were added during earlier commit.  There is no effect on other parts of the code besides the two affected files:
        phys/module_ra_rrtmg_lwf.F
        phys/module_ra_rrtmg_swf.F

2. A change to the Thompson table initialization logic.

This follows up from my most recent commit, that added the capability to
initialize the Thompson microphysics tables from files.  Recall there was
some discussion about changing the namelist controls for this behavior to
make it as streamlined as possible for the average user.  The proposed
change this week implements the following logic for deciding whether to
read or compute the tables:

For each one of three tables:
  Attempt to open and read the table from the associated file.  IF there
  is a failure opening and reading the file:
     IF force_read_thompson is .false. (default) compute the tables
     ELSE  abort.

  IF the tables were computed, above:
     IF write_thompson_tables is .true. (default) write the table file

Basically, if a user does nothing, then the first time they run the
model it will generate the table files in the run directory (assuming
the user hasn't placed them there).  All subsequent runs in that
directory will read the table files that were generated by the first
run in the directory.

If a user wants to make sure that files are read and not computed,
they should change force_read_thompson to .true. in the namelist.

If they always want the tables to be computed and not read, they should
make sure the files don't exist and then set write_thompson_tables to
.false. in the namelist.

The code has also been modified so that only task 0 reads and writes
the table files.

     phys/taug_cpu_defs.h
     phys/taug_cpu_args.h
     phys/module_ra_rrtmg_swf.F
     phys/module_ra_rrtmg_lwf.F
     phys/module_mp_thompson.F
     Registry/Registry.EM_COMMON
     Registry/Registry.NMM




git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@7938 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d