Two changes to make to the WRFDA code, both related to radar assimilation:
First, moving several radar-specific bits of code to their own subroutines; this will allow users to simultaneously assimilate radar data and conventional data in either PREPBUFR or ascii format. Previously, the radar-specific observation reading routines were not accessed when the user is assimilating PREPBUFR obs.

Second, in the process of rearranging the radar-reading code, I'm also removing the broken radar thinning capability which can cause WRFDA to crash if thinning of conventional data is turned on.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@8177 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d