1) Inside of real_em.F and ideal_em.F, put a variable inside of grid that says "I am an ideal case", or "I am not an ideal case".
2) When the output is written (this is the wrfinput_d01 file), this 
logical flag of idealness is used to fill in some metadata.  The 
metadata is important, as that info is read in before anything in the 
Registry.  This metadata is also written during a restart, and for 
a history file.

3) When the WRF model starts, it reads in the metadata to determine 
the idealness of the input data, and makes the decision concerning 
the hypsometric option.  Regardless of the source of the input data 
(wrfinput_d0x or the restart file), the correct info about the ideal 
data is available.

Modify the existing flag this_is_an_ideal_run to be irh.
M       Registry/Registry.EM_COMMON

A one-line mod in each initialization main program, just before the 
wrfinput_d01 write.  Define the value of the LOGICAL variable: 
this_is_an_ideal_run.
M       main/ideal_em.F
M       main/real_em.F

If the metadata exists to determine if this is indeed an ideal run, 
use it (metadata, there is not grid% info yet) to set switches for 
the hypsometric option.  If not, make the same old guess as before, 
based on a date.  For all new data, this is a solid fix.  For old 
existing data, even with this new mod, we are still going to mess up 
restarts with ideal cases with non- year 1 AD starting dates.
M       share/input_wrf.F

Write out the metadata.
M       share/output_wrf.F




git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@8185 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d