H15A-ozone: more changes to fix radiation bugs:
1. The XLAT and XLONG variables are not NMM variables, and are never
initialized by the NMM lat-lon computation routines.  The correct
variables are:
  HLAT,HLON - degrees on H (mass) points
  VLAT,VLON - degrees on V (velocity) points
  GLAT,GLON - radians on H (mass) points
I removed XLAT and XLONG and replaced all references in the code with
HLAT and HLON.

2. Removed the kludgy namelist-based julyr,julday,gmt variable
initialization and replaced it with the better one used in non-HWRF
NMM based on the grid clock.

3. For debugging purposes, set up output of OZMIXM to the main history
stream.

4. The various *aer3d_sw variables are passed in to the radiation
either unallocated, or allocated as 1x1x1x1 variables, but are always
present().  This causes the radiation routines to use those variables
when they shouldn't, and read from random memory locations.  This is
corrected.  This affects ARW and NMM.

5. For debugging purposes, hard-coded levsiz to 59 in the Registry.
This is not needed and should be removed once debugging is done.

6. Added numerous calls to wrf_error_fatal and wrf_debug for debugging
purposes.  These should be removed when debugging is done, along with
any unneeded character*255::message declarations.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/branches/H15A-ozone@8195 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d