File name bug fix for PNETCDF and NOCOLONS:
A change/bug was introduced with -r7980 (2015-01-17) that
modified the output filenames for the PNETCDF I/O form and
also when the NOCOLONS=T flag was set.  This is inconsistent
with the other I/O form options, and so is fixed with this
commit to return the filename behavior to its previous state.

- Remove the .nc 
- Remove the change of dash-to-underscore
- Leave the colon-to-underscore in place for PNETCDF (MPI I/O doesn't like colons)

To Do: consider adding PNETCDF and NOCOLONS test to the regression suite

M    external/io_pnetcdf/wrf_io.F90
M    share/module_io_domain.F


git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@8406 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d