Two requested changes by an AIX user for the build:
1) Remove the -tradtional flag on the cpp command.

The purpose of the -traditional flag is to allow cpp to
work with Fortran codes.  Such things that would not be
proper in C or C++ (single apostrophes to use the possessive
in comments) or interpreted as comments (the Fortran 
syntax for concatentation "//") are usually handled by the
-traditional flag.  The user says use of this cpp flag causes 
troubles.  For this one AIX stanza we are removing the 
"-traditional" by commenting it out in the Makefile TRADITIONAL
macro.

2) Increase the buffer size for m4 (from -B 14000 to -B 20000).

For our easily availble machines to test (Mac desktop and 
yellowstone), this seems to be a NOOP based on the output from
m4 processing.

For example, here is a sample text file: foo1
ABC
ABCD

The expected results from this command 
m4 -DABC=foo foo1
are:
foo
ABCD

By adding a buffer size: no changes AND a note that this is likely
going away.

m4 -B 1 -DABC=foo foo1
m4: Warning: `m4 -B' may be removed in a future release
foo
ABCD

m4 -B 5000 -DABC=foo foo1
m4: Warning: `m4 -B' may be removed in a future release
foo
ABCD

There is no -B option list in the any man page I can find.  Therefore,
modifying the -B buffer size is unlikely to cause bad results for most
users, and will help users on this type of machine that is failing with
the -B buffer too small.

M    configure_new.defaults


git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@8419 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d