Changes related to io_netcdf directory paths and remove PIO build target.
Small changes, with no impact on the WRF build, but these
get rid of annoying (and unimportant Errors) listed in the build log
(which have users sending email to wrfhelp).

1) When already in the WRFV3/eternal/io_netcdf directory to build diffwrf,
to build the libraries in io_netcdf (the second trip in), the cd command
is wrong
-          cd ../external/io_netcdf ; \
+          cd ../io_netcdf ; \
This command works accidentally, because the cd command fails, and the 
Makefile stays in the same directory.

2) Until the full-blown PIO is building correctly, let us remove the build
target in the top level Makefile.

3) There were a few commented out lines, those are just deleted.

M    Makefile


git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@8421 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d