TYPE: release
KEYWORDS: release version

SOURCE: internal

PURPOSE: update version_decl string from 3.7 to 3.7+ to clear up release is post 3.7

DESCRIPTION OF CHANGES: The WRF model relies on the string "release_version" when printing out the initial model output.
This is a 10-character string that is kept updated with the current version of the model source code.  It has been updated
from
"V3.7      "
to
"V3.7+     "

LIST OF MODIFIED FILES (annotated if not obvious, not required to be on a single line):
M       inc/version_decl

TESTS CONDUCTED:
1. The WRF code compiles on my laptop with GNU 4.8.1.
2. WRF and ideal print out the expected string at the top of the standard print-out.


git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@8467 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d