KEYWORDS: WRF, IO, stream identifier
SOURCE: NCEP
 
PURPOSE: Fix off-by-one error in WRF I/O code dealing with stream assignments
 
DESCRIPTION OF CHANGES: 
Gentle developers,
 
I'm proposing a bug fix for an off-by-one error in a section of code
that works out what variable-set goes on WRF input or output streams.

Error surfaced when 202 was used in HWRF code to indicate a stream
that was not supposed to be subject to quilting, even when quilt I/O
was enabled for other streams.  The off-by-one error caused the stream
in question to be misidentified as AUXHIST0 in the mediation_integrate
logic and, since there is no such stream, the variable set on the stream
defaulted to HISTORY and putting the quilt servers and their clients
out of sync for this stream -- resulting in a hang.  The hang would
not appear under the current WRF code, but the new nest parallelization
scheme in HWRF (set to be proposed for WRF trunk later this year) has
a different organization of quilt servers to clients.

It's been in the code for a long time so that one is  (1) surprised that
the bug has never shown up before and (2) worried that changing it may
create some ripples, if there have been other changes over the years
to fix some behavior that was noticed but not attributed to this error
(in other words, fixing this original bug might break the workarounds
in WRF or other codes that use WRF I/O).

The error is in the two routines open_hist_w and open_aux_u that figure
out, based on the stream number, which kind of stream it is.   But the
arithmetic has an off-by one error.  The result has been that HISTORY
datasets have been misnamed AUXHIST0 by this routine, which isn't a
valid stream designator.  But that error is ignored because there's
logic  elsewhere that just calls any stream that doesn't have a valid
identifier history:

    ELSE  ! default if nothing is set in SysDepInfo; use history
      CALL nl_get_io_form_history( 1, io_form )
    ENDIF

One may have noticed lots of warning messages with NetCDF complaints about
missing fields when reading some file when debug_level is greater than
0 in the namelist.  One suspects this bug might be the reason.    The bug
could also have led to some head-scratchers over the years with the
AUXHIST and AUXINPUT streams not working as advertised, but it would
be infrequent and weird enough that people might just have thought,
"just WRF being WRF".

LIST OF MODIFIED FILES :    share/mediation_integrate.F




git-svn-id: https://svn-wrf-model.cgd.ucar.edu/branches/RELEASE_BRANCH_371@8478 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d