Brought over from the 3.7.1 Bugfix release Original commit message: ------------------------------------------------------------------------ r8474 | dudhia | 2015-07-10 10:26:00 -0600 (Fri, 10 Jul 2015) | 21 lines
TYPE: bug-fix
 
KEYWORDS: Revised surface layer scheme out-of-bounds issue, computed index
 
SOURCE:  NCAR
 
PURPOSE: Fix out of bounds problem in look-up tables
 
DESCRIPTION OF CHANGES: One line added to ensure z/L guess in iteration is not
opposite sign to Ri as this goes out of bounds, limit z/L to zero.
      if(zol2*ri2 .lt. 0.)zol2=0.  ! limit zol2 - must be same sign as ri2
Note that this is not a frequent occurrence, and that the iteration does
eventually get to the right answer, but it is not clear what the out-of-bounds
value will produce.

LIST OF MODIFIED FILES :  phys/module_sf_sfclayrev.F
 
TESTS CONDUCTED:   Jan 2000 case exhibited this only in first time step, and
at one point, and this was resolved. User has also tested this with success.

------------------------------------------------------------------------


git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@8599 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d