Brought over from the 3.7.1 Bugfix release Original commit message: ------------------------------------------------------------------------ r8515 | weiwang | 2015-07-31 16:50:28 -0600 (Fri, 31 Jul 2015) | 23 lines
TYPE: bug fixes

KEYWORDS: diffuse_frac, Goddard radiation, NaNs, Registry

SOURCE:  Internal

PURPOSE: Fix NaNs in DIFFUSE_FRAC calculation (when Goddard radiation is used).

DESCRIPTION OF CHANGES: The DIFFUSE_FRAC calculation in radiation driver checks
for coszen > 1e-3, but when Goddard radiation is used, SWDOWN can be zero
with small coszen value. The change is to check SWDOWN directly since that
is used as denominator. DIFFUSE_FRAC is also checked after the calcuation
to limit it between 0 and 1. As this field does not need to be present in
the history file, it is now removed.

LIST OF MODIFIED FILES :
   phys/module_radiation_driver.F
   Registry/Registry.EM_COMMON

TESTS CONDUCTED: WTF passed on release branch. Calculation checked before 
and after the change. NaNs removed.


------------------------------------------------------------------------




git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@8623 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d