====================================== TYPE: bug fix
KEYWORDS: WRFDA, OBSPROC, 

SOURCE: internal

PURPOSE: Fix incomplete deallocation/disassociation of linked lists in OBSPROC, which can cause segmentation faults in certain situations due to double-calling of the deallocation subroutine.

DESCRIPTION OF CHANGES: 
OBSPROC reads LITTLE_R format observation files and converts them into a text format which can be read by WRFDA. If a LITTLE_R file is written incorrectly (for example: some fields can be written as "***********" instead of the proper missing data value if the program used to write them does not properly deal with abnormal observation values; this is due to how Fortran deals with incorrect write format statements), OBSPROC has routines to deal with bad data reads. However, for the specific error case where there is a bad read for measurement values in the data records of a multi-level observation, there can be segmentation faults due to a logic error in the current version of OBSPROC: 

OBSPROC uses the subroutine "dealloc_meas" to deallocate multi-level observations which are stored as linked lists. This subroutine deallocates the first measurement, then follows the pointer to the next measurement, deallocates that, etc. However, "dealloc_meas" fails to take into account that the head of this linked list is stored as a pointer within the larger observation "report" data structure, which contains time, location, and other information. At the very end of the subroutine, it fails to nullify the initial pointer to the head of the linked list of measurement levels. This means that if OBSPROC checks later to see if this measurement has been deallocated yet (by checking if the head pointer is associated), it will think that it has not yet been deallocated, and call "dealloc_meas" again. Since these arrays actually *have* been deallocated already, attempting to deallocate them again will usually result in a segmentation fault.

The solution is a one-line change: adding a statement to nullify the head pointer at the end of the "dealloc_meas" subroutine.

LIST OF MODIFIED FILES :
M       var/obsproc/src/module_decoded.F90

TESTS CONDUCTED: Inserted a bad observation into the input LITTLE_R data for OBSPROC regression tests, which caused them to segfault with the trunk code. Applying the fix fixes the segfault, and the results are bit-for-bit identical to previous test results.
=======================================



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@8678 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d