======================================= TYPE: bug fix
KEYWORDS: Registry, missing quote

SOURCE: internal

PURPOSE: A quote was missing for the state variable PTROPNN, at the end of the description string.

DESCRIPTION OF CHANGES: 
This is one of those one character changes.  A quote is added to close the character string for the variable description.

-state real ptropnn ij  dyn_em  1  -  i1  "PTROPNN"  "PTROP, nearest neighbor interp      "Pa"
+state real ptropnn ij  dyn_em  1  -  i1  "PTROPNN"  "PTROP, nearest neighbor interp"     "Pa"
                                                                                    /\
                                                                                    |
NOTE: ------------------------------------------------------------------------------|

LIST OF MODIFIED FILES (annotated if not obvious, not required to be on a single line): 
M	Registry/Registry.EM_COMMON

TESTS CONDUCTED:
Several users mentioned this missing quote causing troubles, but we could not replicate that error.
The regression test performs as expected.
=======================================


git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@8697 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d