======================================= TYPE: new feature
KEYWORDS: advection, timing test, main program, wrapper, cpp, ADVECT_KERNEL

SOURCE: internal

PURPOSE: Put an ifedf'ed main program into the advection model to allow consistent timing tests to be conducted with an always
up-to-date version of the advection scheme.  Right now, the only scheme tested is positive definite scalar advection.  The
looping structure is consistent in all of the advection routines (and based on tests, that looping and the resultant cache misses
are the troubles).  As cache misses are the issue, this is set up to build serially only.

To activate, the following is in the code:
!  cpp -traditional -C -P -DADVECT_KERNEL module_advect_em.F > advection_kernel.f90
!  ifort, pgfortran, gfortran advection_kernel.f90
!  a.out

DESCRIPTION OF CHANGES:
Using #ifdefs, a driver main program that calls the advect_scalar_pd scheme is inserted into the advection module
file.  Unless the user specifies ADVECT_KERNEL on the cpp command, the main program is not included in the source
code that gets compiled.

LIST OF MODIFIED FILES:
M       dyn_em/module_advect_em.F

TESTS CONDUCTED:
1. regression test OK (no changes expected or seen in results)
2. main program is built (when compiled), and would be eligible to give timing results
Example output:
 LOOP over scalars
 CALL advect for loop =            1 , of           15  loops
 CALL advect for loop =            2 , of           15  loops
 CALL advect for loop =            3 , of           15  loops
 CALL advect for loop =            4 , of           15  loops
 CALL advect for loop =            5 , of           15  loops
 CALL advect for loop =            6 , of           15  loops
 CALL advect for loop =            7 , of           15  loops
 CALL advect for loop =            8 , of           15  loops
 CALL advect for loop =            9 , of           15  loops
 CALL advect for loop =           10 , of           15  loops
 CALL advect for loop =           11 , of           15  loops
 CALL advect for loop =           12 , of           15  loops
 CALL advect for loop =           13 , of           15  loops
 CALL advect for loop =           14 , of           15  loops
 CALL advect for loop =           15 , of           15  loops
=======================================


git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@8703 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d