TYPE: bug fix
KEYWORDS: check_a_mundo, use_wps_input, if test

SOURCE: internal

PURPOSE: A new IF test was introduced incorrectly into the real program's portion of check_a_mundo.  The flag to be tested is an integer.  The IF test syntax assumed the value was a logical.  This was picked up by GNU compilers.

DESCRIPTION OF CHANGES: 
The IF test is supposed to be executed when the real program is the calling routine.  That is when use_wps_input==1.

Index: share/module_check_a_mundo.F
===================================================================
--- share/module_check_a_mundo.F	(revision 8704)
+++ share/module_check_a_mundo.F	(working copy)
@@ -1143,7 +1143,7 @@
 !     lowest level from surface option to be activated.
 !-----------------------------------------------------------------------
 
-      IF ( model_config_rec % use_wps_input ) THEN
+      IF ( model_config_rec % use_wps_input .EQ. 1 ) THEN
          IF ( ( .NOT. model_config_rec % use_surface )  .AND. &
               ( model_config_rec % force_sfc_in_vinterp .GT. 0 ) ) THEN
             wrf_err_message = '--- NOTE: Inconsistent vertical interpolation settings in program real.'


LIST OF MODIFIED FILES (annotated if not obvious, not required to be on a single line): 
M       share/module_check_a_mundo.F

TESTS CONDUCTED:
Compiles GNU successfully.
Passes Intel, PGI, GNU regtests.


git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@8705 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d