TYPE: bug fix, SIGNIFICANT IMPACT
KEYWORDS: icloud = 3, reset qi and qc

SOURCE: Greg Thompson

PURPOSE: reset qc and qi to their microphysics-only values when using ICLOUD=3 option

DESCRIPTION OF CHANGES: 
When icloud=3 option is used, the temporary values of cloud water and ice are changed to include estimates from the RH-based partial cloud scheme in radiation driver. After the radiation is called, these fields need to be reset to the original, microphysics-only values. But they were not reset in the original implementation. This fix corrects the problem, and removes a large cold bias seen in the initial implementation. 

LIST OF MODIFIED FILES 

phys/module_radiation_driver.F

TESTS CONDUCTED:

Standard Jan 2000 case, and reg test. A 20 km CONUS case from June 11 2010 was used to look at the effect of the change.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@8754 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d