====================================== TYPE: bug fix
KEYWORDS: WRFDA, hybrid, vertical localization

SOURCE: internal

PURPOSE: Fix segmentation fault within executable gen_be_vertloc.exe

DESCRIPTION OF CHANGES:

The executable gen_be_vertoc.exe creates a file 'be.vertloc.dat', which is used for vertical localization in WRFDA hybrid 3dvar/ensemble assimilation. An old bit of code in this program has an incorrectly hard-coded value for the variable "ktarget", which results in writing beyond the bounds of an array if the number of vertical levels is smaller than that hard-coded value. This results in a segmentation fault at the deallocation stage of the program.

Even though the error occurs after the file is successfully written, it can cause an indefinite hang on Yellowstone, so it needs to be fixed. Simply correcting the hard-coded value to be equal to kz (the number of vertical levels) fixes this problem.

LIST OF MODIFIED FILES :
M       WRFDA/var/gen_be/gen_be_vertloc.f90


TESTS CONDUCTED: The fix results in successful completion of gen_be_vertloc.exe and successful output of be.vertloc.dat from that program, which is then used successfully in the new hybrid regression test.

=======================================



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@8796 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d