TYPE: enhancement
KEY WORDS: Snow fraction

SOURCE: developed at ESRL as part of RAP/HRRR development

PURPOSE: adds the capability to compute snow fraction as a function of land-use types.

DESCRIPTION OF CHANGES:
Snow fraction with the forest land-use is allowed to be two times smaller that for grassland/cropland, etc. with the same snow depth. Limit urban snow fraction to be less than 75% - urban category variable is added to the VEGPARM.TBL for this enhancement. Section for NLCD is moved ahead of RUC LSM because of added new line for hard-wired read. Section for NLCD in VEGPARM.TBL is moved ahead of RUC LSM because of added new line for hard-wired read.

LIST OF MODIFIED FILES

M  phys/module_sf_ruclsm.F
M  run/VEGPARM.TBL

TESTS CONDUCTED:

Tested in RAP and HRRR on different number of processes with identical results. Reg tested.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@8944 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d