TYPE: bug fix
KEYWORDS: GNU DFI time step timestep

SOURCE: internal

PURPOSE: incorrect 1 second time step with DFI simulations with GNU

DESCRIPTION OF CHANGES: 
The default value for the time_step_dfi was a "-" character in the Registry.EM_COMMON.  This has been changed to a "-1", which is more appropriate for an integer.  Also, tests inside check_a_mundo.F now detect the flag value.  If a time step of -1 is found with running any of the DFI steps, then the time step defaults to the standard WRF time step.

LIST OF MODIFIED FILES (annotated if not obvious, not required to be on a single line): 
M       Registry/Registry.EM_COMMON
M       share/module_check_a_mundo.F

TESTS CONDUCTED (explicitly state mandatory, voluntary, and assigned tests, not required to be on a single line):
1) Regression tests for GNU now run correctly.
2) Regression tests include namelists with and without explicit time_step_dfi, so the test is in both cases working.


git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@9008 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d