TYPE: bug fix
KEYWORDS: double replicated printout

SOURCE: John Michalakes (NOAA)

PURPOSE: remove instances of duplicated print from the WRF system stdout and stderr

DESCRIPTION OF CHANGES: 
A modification was introduced to reduce the stdout and stderr logging for the
HWRF package.  The control of that in the WRF system is now correctly modified
to allow ARW serial codes to NOT turn on the stderr_logging.

LIST OF MODIFIED FILES (annotated if not obvious, not required to be on a single line): 
M	frame/module_wrf_error.F

TESTS CONDUCTED (explicitly state mandatory, voluntary, and assigned tests, not required to be on a single line):
1. No double printing of stdout after this fix.



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@9019 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d