TYPE: bug fix
KEYWORDS: module_io bounds string

SOURCE: internal

PURPOSE: Fix an out-of-bounds string reference.  This bounds error was likely introduced when the maximum size of the file name lengths were recently increased from 80 to 256 characters.

DESCRIPTION OF CHANGES: 
In subroutine get_value_from_pairs, replace a declaration for a local character string from 128 to 256 (where 256 is the new length of allowable file names).

LIST OF MODIFIED FILES (annotated if not obvious, not required to be on a single line): 
M	frame/module_io.F

TESTS CONDUCTED (explicitly state mandatory, voluntary, and assigned tests, not required to be on a single line):
1) Regression test - do no harm
2) Passes bounds check with the fix, and does not pass without the fix.


Index: module_io.F
===================================================================
--- module_io.F	(revision 9069)
+++ module_io.F	(working copy)
@@ -2626,7 +2626,7 @@
  CHARACTER*(*) ::    str
  CHARACTER*(*) ::    retval

-  CHARACTER (128) varstr, tstr
+  CHARACTER (256) varstr, tstr
  INTEGER i,j,n,varstrn
  LOGICAL nobreak, nobreakouter



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@9078 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d