KEYWORDS: rogue comma Cray
SOURCE: Patricia Balle (Cray)

PURPOSE: Remove syntactically incorrect comma in an internal write statement.

DESCRIPTION OF CHANGES: 
Original code looks something like this:
WRITE ( string , format ) , list_of_variables
It should, instead, look like this:
WRITE ( string , format )   list_of_variables
The line in question is quite complicated with all sorts of single and double quotes and 8 commas that should be there, and the extra comma was just not really noticed.  The Intel, PGI, and GNU compilers do not mind the additional comma.

LIST OF MODIFIED FILES (annotated if not obvious, not required to be on a single line): 
M	external/RSL_LITE/module_dm.F

TESTS CONDUCTED (explicitly state mandatory, voluntary, and assigned tests, not required to be on a single line):
1) Regression test OK - do no harm.

2) Easily available online standard (which agrees with the Fortran 2003 text):

https://www.fortran.com/F77_std/rjcnf0001-sh-12.html

12.8 READ, WRITE, and PRINT Statements

The READ statement is the data transfer input statement. The WRITE and PRINT statements are the data transfer output statements. The forms of the data transfer input/output statements are:
READ (cilist) [iolist]
READ f [,iolist]
WRITE (cilist) [iolist]         <<<<<---- This is the one we are hanging our hat on
PRINT f [,iolist]
where: cilist is a control information list ( 12.8.1) that includes:
(1)A reference to the source or destination of the data to be transferred
(2)Optional specification of= editing processes
(3)Optional specifiers that determine the execution sequence on the occurrence of certain events
(4)Optional specification to identify a record
(5)Optional specification to provide t|%he return of the input/output status
f is a format identifier ( 12.4)
iolist is an input/output list ( 12.8.2) specifying the data to be transferred

Index: external/RSL_LITE/module_dm.F
===================================================================
--- external/RSL_LITE/module_dm.F       (revision 9069)
+++ external/RSL_LITE/module_dm.F       (working copy)
@@ -1869,7 +1869,7 @@
         ELSE
           IF ( comm_start(nest_id) .LT. comm_start(parent_id(nest_id)) ) THEN
             WRITE(wrf_err_message,&
-        "('nest domain ',i3,'comm_start (',i3,') lt parent ',i3,' comm_start (',i3,')')"),&
+        "('nest domain ',i3,'comm_start (',i3,') lt parent ',i3,' comm_start (',i3,')')") &
                    nest_id,comm_start,parent_id(nest_id),comm_start(parent_id(nest_id))
             CALL wrf_error_fatal(TRIM(wrf_err_message))
           ELSE IF ( comm_start(nest_id) .LT. &



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@9082 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d